[APINotes] NFC: extract a test for namespaces
rdar://113403829
