+++ clang/test/APINotes/Inputs/Frameworks/CXXInteropKit.framework/Headers/CXXInteropKit.apinotes
@@ -10,42 +10,6 @@ Classes:
 Enumerators:
 - Name: SomeClassRed
   SwiftName: red
-Functions:
-- Name: funcInNamespace
-  SwiftName: inWrongContext()
 Tags:
 - Name: NSSomeEnumOptions
   SwiftName: SomeEnum.Options
-- Name: char_box
-  SwiftName: InWrongContext
-Namespaces:
-- Name: Namespace1
-  Globals:
-  - Name: varInNamespace
-    SwiftName: swiftVarInNamespace
-  Functions:
-  - Name: funcInNamespace
-    SwiftName: swiftFuncInNamespace()
-  Tags:
-  - Name: char_box
-    SwiftName: CharBox
-  Namespaces:
-  - Name: Nested1
-    Globals:
-    - Name: varInNestedNamespace
-      SwiftName: swiftVarInNestedNamespace
-    Functions:
-    - Name: funcInNestedNamespace
-      SwiftName: swiftFuncInNestedNamespace(_:)
-    Tags:
-    - Name: char_box
-      SwiftName: NestedCharBox
-    Namespaces:
-    - Name: Namespace1
-      Tags:
-      - Name: char_box
-        SwiftName: DeepNestedCharBox
-  - Name: Nested2
-    Globals:
-    - Name: varInNestedNamespace
-      SwiftName: swiftAnotherVarInNestedNamespace
