+++ clang/test/APINotes/Inputs/Headers/module.modulemap
@@ -21,3 +21,7 @@ module ModuleWithWrongCase {
 module ModuleWithWrongCasePrivate {
   header "ModuleWithWrongCasePrivate.h"
 }
+
+module Namespaces {
+  header "Namespaces.h"
+}
